renamed HandGenerator to TestHandGenerator
